Create iterator over player messages
