+++ backend/tests/test_game_state.py
@@ -1,0 +1,48 @@
+from backend.channel import channel
+from backend.game_state import Lobby, Player
+from backend.models import Role
+
+
+def test_lobby_broadcast() -> None:
+    """Broadcasted messages are received from multiple players."""
+    left_a, right_a = channel()
+    left_b, right_b = channel()
+
+    lobby = Lobby({"A": Player(left_a, Role.manager), "B": Player(left_b, Role.burger)})
+
+    msg = 5
+    lobby.broadcast(msg)  # pyright: ignore[reportArgumentType]
+
+    assert right_a.get() == msg
+    assert right_b.get() == msg
+
+
+def test_messages_receives_msg() -> None:
+    """Test iterator over multiple players."""
+    left_a, right_a = channel()
+    left_b, right_b = channel()
+
+    lobby = Lobby({"A": Player(left_a, Role.manager), "B": Player(left_b, Role.burger)})
+
+    right_a.put(4)
+    right_b.put(8)
+
+    messages = iter(lobby.messages())
+    assert next(messages) == 4
+    assert next(messages) == 8
+
+
+def test_messages_skips_none() -> None:
+    """Test that the iterator skips over empty player message channels."""
+    left_a, right_a = channel()
+    left_b, _ = channel()
+
+    lobby = Lobby({"A": Player(left_a, Role.manager), "B": Player(left_b, Role.burger)})
+
+    # two messages from A: we should get both without any Nones in between
+    right_a.put(4)
+    right_a.put(8)
+
+    messages = iter(lobby.messages())
+    assert next(messages) == 4
+    assert next(messages) == 8
